Fix the problem introduced by commit caeee93 that the introduction of condition check (modObj instanceof IProject) causes failures when dragging java project from package explorer to application viewer.
The model of Java Project in eclipse is the instance of IJavaProject, so the original code in line 192 of ApplicationMasterPart can't recognize the DragAndDrop JavaProject to applicationsViewer.

This bug was introduced by a commit caeee93bc4d00c7c413a6da2d3349b2ede780537 for 'Merged IBM Patch for Run On Server fix for module URL navigation' in 2014-03-37.

Current commited modifications can fix this bug, make the applicationsViewer identify the Java Project and presents the Application Push Wizard corrently.